HWKAPM-469 Rename BTxn completion time calculatoin mechanism to Trace completion time - as not specific to named transactions
