+++ processors/tracecompletiontime-deriver/src/main/resources/META-INF/beans.xml
