--- server/api/src/main/java/org/hawkular/apm/server/api/services/BTxnCompletionTimePublisher.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.apm.server.api.services;
-
-import org.hawkular.apm.api.model.events.CompletionTime;
-import org.hawkular.apm.api.services.Publisher;
-
-/**
- * This interface provides the capability for publishing completion time
- * information for end to end business transactions.
- *
- * @author gbrown
- */
-public interface BTxnCompletionTimePublisher extends Publisher<CompletionTime> {
-
-}
