+++ server/cassandra/src/test/java/org/hawkular/apm/server/cassandra/ServicesCassandraTest.java
@@ -846,7 +846,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -877,7 +877,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -915,7 +915,7 @@ public class ServicesCassandraTest {
         cts.add(ct3);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -947,7 +947,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -985,7 +985,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1044,7 +1044,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1104,7 +1104,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1166,7 +1166,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1230,7 +1230,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1285,7 +1285,7 @@ public class ServicesCassandraTest {
         cts.add(ct1_2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(2000);
@@ -1329,7 +1329,7 @@ public class ServicesCassandraTest {
         cts.add(ct1_2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(2000);
@@ -1379,7 +1379,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
@@ -1426,7 +1426,7 @@ public class ServicesCassandraTest {
         cts.add(ct2);
 
         try {
-            analytics.storeBTxnCompletionTimes(null, cts);
+            analytics.storeTraceCompletionTimes(null, cts);
 
             synchronized (this) {
                 wait(1000);
