--- server/jms/src/main/java/org/hawkular/apm/server/jms/BTxnCompletionTimeDeriverMDB.java
@@ -1,69 +1,0 @@
-/*
- * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.apm.server.jms;
-
-import javax.annotation.PostConstruct;
-import javax.ejb.ActivationConfigProperty;
-import javax.ejb.MessageDriven;
-import javax.ejb.TransactionAttribute;
-import javax.ejb.TransactionAttributeType;
-import javax.ejb.TransactionManagement;
-import javax.ejb.TransactionManagementType;
-import javax.inject.Inject;
-import javax.jms.MessageListener;
-
-import org.hawkular.apm.api.model.events.CompletionTime;
-import org.hawkular.apm.processor.btxncompletiontime.BTxnCompletionInformation;
-import org.hawkular.apm.processor.btxncompletiontime.BTxnCompletionTimeDeriver;
-import org.hawkular.apm.server.api.services.BTxnCompletionTimePublisher;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-
-/**
- * @author gbrown
- */
-@MessageDriven(name = "BTxnCompletionInformation_BTxnCompletionTimeDeriver",
-        messageListenerInterface = MessageListener.class,
-        activationConfig =
-        {
-                @ActivationConfigProperty(propertyName = "destinationType", propertyValue = "javax.jms.Topic"),
-                @ActivationConfigProperty(propertyName = "destination", propertyValue = "BTxnCompletionInformation"),
-                @ActivationConfigProperty(propertyName = "subscriptionDurability", propertyValue = "Durable"),
-                @ActivationConfigProperty(propertyName = "clientID", propertyValue = "BTxnCompletionTimeDeriver"),
-                @ActivationConfigProperty(propertyName = "subscriptionName",
-                            propertyValue = "BTxnCompletionTimeDeriver")
-        })
-@TransactionManagement(value = TransactionManagementType.CONTAINER)
-@TransactionAttribute(value = TransactionAttributeType.REQUIRED)
-public class BTxnCompletionTimeDeriverMDB extends ProcessorMDB<BTxnCompletionInformation, CompletionTime> {
-
-    @Inject
-    private BTxnCompletionInformationPublisherJMS btxnCompletionInformationPublisher;
-
-    @Inject
-    private BTxnCompletionTimePublisher btxnCompletionTimePublisher;
-
-    @PostConstruct
-    public void init() {
-        setProcessor(new BTxnCompletionTimeDeriver());
-        setRetryPublisher(btxnCompletionInformationPublisher);
-        setPublisher(btxnCompletionTimePublisher);
-        setTypeReference(new TypeReference<java.util.List<BTxnCompletionInformation>>() {
-        });
-    }
-
-}
